tooltip improve
