+++ src/pages/Transaction/Tabs/Components/TransactionFunction.tsx
@@ -1,58 +1,11 @@
 import React from "react";
-import {Box, Stack, SxProps, Theme, useTheme} from "@mui/material";
+import {Box, Stack, SxProps, Theme} from "@mui/material";
 import {Types} from "aptos";
 import CurrencyExchangeOutlinedIcon from "@mui/icons-material/CurrencyExchangeOutlined";
 import DescriptionOutlinedIcon from "@mui/icons-material/DescriptionOutlined";
+import {CodeLineBox} from "../../../../components/CodeLineBox";
 
-const TEXT_COLOR_LIGHT = "#0EA5E9";
-const TEXT_COLOR_DARK = "#83CCED";
-const BACKGROUND_COLOR = "rgba(14,165,233,0.1)";
 const BLOCK_MODULE_NAME = "candy_machine_v2";
-
-export function CodeLineBox({
-  children,
-  sx,
-}: {
-  children: React.ReactNode;
-  sx?: SxProps<Theme>;
-}) {
-  const theme = useTheme();
-  return (
-    <Box
-      sx={[
-        {
-          width: "max-content",
-          color:
-            theme.palette.mode === "dark" ? TEXT_COLOR_DARK : TEXT_COLOR_LIGHT,
-          backgroundColor: BACKGROUND_COLOR,
-          padding: "0.35rem 1rem 0.35rem 1rem",
-          overflow: "hidden",
-          whiteSpace: "nowrap",
-          textOverflow: "ellipsis",
-          fontFamily: theme.typography.fontFamily,
-          fontWeight: theme.typography.fontWeightRegular,
-          fontSize: 13,
-        },
-        ...(Array.isArray(sx) ? sx : [sx]),
-      ]}
-      borderRadius={1}
-    >
-      {children}
-    </Box>
-  );
-}
-
-function CodeLine({
-  data,
-  sx,
-}: {
-  data: string;
-  sx?: SxProps<Theme>;
-}): JSX.Element {
-  return (
-    <CodeLineBox sx={[...(Array.isArray(sx) ? sx : [sx])]}>{data}</CodeLineBox>
-  );
-}
 
 function CoinTransferCodeLine({sx}: {sx?: SxProps<Theme>}): JSX.Element {
   return (
@@ -112,6 +65,8 @@ export default function TransactionFunction({
   }
 
   return (
-    <CodeLine data={functionStr} sx={[...(Array.isArray(sx) ? sx : [sx])]} />
+    <CodeLineBox sx={[...(Array.isArray(sx) ? sx : [sx])]}>
+      {functionStr}
+    </CodeLineBox>
   );
 }
