+++ src/pages/Validators/Components/RewardsPerformanceTooltip.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import {Stack, Typography} from "@mui/material";
 import TableTooltip from "../../../components/Table/TableTooltip";
-import {CodeLineBox} from "../../Transaction/Tabs/Components/TransactionFunction";
+import {CodeLineBox} from "../../../components/CodeLineBox";
 
 export default function RewardsPerformanceTooltip() {
   return (
@@ -18,7 +18,7 @@ export default function RewardsPerformanceTooltip() {
           sx={{
             whiteSpace: "normal",
             wordWrap: "break-word",
-            fontSize: "inherit",
+            fontSize: 13,
             padding: "0.3rem 0.5rem 0.3rem 0.5rem",
             width: "100%",
           }}
